Add tests for SDK containers.
Prior to this change, an integration test suite only operated on the
"kitchen sink" pulumi/pulumi image which contains all 4 supported SDKs
for Pulumi.  This change adds the ability to run the tests for a subset
of SDKs, along with actually running the appropriate test suite for the
UBI SDK-specific Docker images during the CI build.

(Debian images build for multiple platforms and are a little tricker -
they will be addressed in a future PR.)