Fix golang image.
The PULUMI_VERSION arg was defined before the FROM directive in the
build container.  We believe this caused "latest" to be the value read
by the line that installs Pulumi, and thus led to a mismatch between a
tagged image and the version of Pulumi installed on the image.
(get.pulumi.com does not update the latest version until after the docs
are published, but the Docker images' release build fires before the
docs are updated.)