Bump NodeJS in base image to 18 (#150)
* Bump NodeJS in base image to NodeJS LTS (v18)

* Use `node:18-bullseye-slim` for the `pulumi/nodejs` image

Before it was using `lts-bullseye-slim`, which is currently the same as `18-bullseye-slim`. This commit changes it it to `18-bullseye-slim` so that it remains on 18 once the LTS version changes. That way, we can upgrade all images to the next LTS version at the same time, and not have one images that is using a different version that the others.

* Upgrade the `pulumi/nodejs` `-ubi` image to use Node 18

* Update CHANGELOG.md

* Update README.md with policy

Also tweak the cadence as we aim to release weekly now.

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>