Matrix Python debian images (#226)
Create per language version images for python 3.9 to 3.12. The image
without suffix `pulumi/pulumi-python` continues to use 3.9.

Other languages or UBI based images are not matrixed, but this PR puts
in place most of the infrastructure for this. To generate the matrix, we
use a python script that creates a list of `include` entries for the
matrix. We do this instead of using the normal way to specify a matrix
using variables because each language has its own range of versions.

For example for Python 3.12 we get the tags/manifests:
* pulumi-python-3.12:latest
* pulumi-python-3.12:3.127.0
* pulumi-python-3.12:3.127.0-debian
* pulumi-python-3.12:3.127.0-debian-arm64
* pulumi-python-3.12:3.127.0-debian-amd64

Example on my docker hub:
https://hub.docker.com/repository/docker/jpoissonnier/pulumi-python-3.12/tags

```
docker run --rm -it --entrypoint python jpoissonnier/pulumi-python-3.12:latest --version
Python 3.12.4
```