Update pulumi dockerfile to multiarch (#274)
I would like to setup devcontainers for use by the developers when they
are working on my custom provider. Currently since there isn't an arm64
version, the experience is slow and painful on Apple Silicon. This
change updates the Dockerfile so that all that is needed is the
`--platform=linux/arm64` in the build and it will take care of switching
out all the places needed to build. Note: The build matrices will need
to be updated if this is something you would consider adding.

Thanks.

---------

Co-authored-by: Julien <julien@caffeine.lu>