Pin Gradle to 8.14.2 and Maven to 3.9.10 in the UBI Java image (#487)
For the UBI image we use [SDKMAN!](https://sdkman.io) to install gradle
and maven.

Due to https://github.com/sdkman/sdkman-cli/issues/1458 we have to
explicitly ask for Maven 3.9.10, otherwise 3.9.9 will be installed. Pin
Gradle to the latest version while we are at it. This better matches the
behaviour of the Debian based images, which install a predictable
version (what’s in the repo).