Catch errors in tests of the kitchen sink image (#495)
While looking into https://github.com/pulumi/templates/pull/929 I
noticed that the kitchen sink job was not failing
https://github.com/pulumi/pulumi-docker-containers/actions/runs/15796277766/job/44529555986#step:16:1064
when it should have run into the same bug. The tests actually do fail,
but our setup hides that from CI.

To fix this we run a matrix on the variant: normal and non root.

Also fixes the image name for the `nonroot` variant.