Fix tag fetching (#525)
We need to also set fetch-depths: 0 to fetch everything, along with
fetch-tags to get the tags. The repo currently has just under 500
commits, so that should be fine. If we are concerned about the checkout,
we could also fetch the last 100 commits or so. We won't have 100
commits (with squash committing) in a release.